{"schema":"decodex/commit/1","summary":"Split MCP server modules","authority":"XY-1115"}
