+++ apps/decodex/src/mcp/server/stdio.rs
@@ -1,0 +1,53 @@
+use std::io::{BufRead as _, BufReader, ErrorKind, Read, Write};
+
+use serde_json::{self, Value};
+
+use crate::{
+	mcp::{McpCapabilityProfile, McpContext, McpTransport, server::core::McpServer},
+	prelude::Result,
+};
+
+pub(in crate::mcp) fn serve_stdio_with_profile<R, W>(
+	reader: R,
+	mut writer: W,
+	context: McpContext,
+	capability_profile: McpCapabilityProfile,
+) -> Result<()>
+where
+	R: Read,
+	W: Write,
+{
+	let server = McpServer { context, capability_profile, transport: McpTransport::Stdio };
+	let reader = BufReader::new(reader);
+
+	for line in reader.lines() {
+		let line = line?;
+
+		if line.trim().is_empty() {
+			continue;
+		}
+
+		for response in server.handle_line(&line, true) {
+			write_json_line(&mut writer, &response)?;
+		}
+	}
+
+	Ok(())
+}
+
+fn write_json_line<W>(writer: &mut W, value: &Value) -> Result<()>
+where
+	W: Write,
+{
+	let line = serde_json::to_string(value)?;
+
+	match writer
+		.write_all(line.as_bytes())
+		.and_then(|()| writer.write_all(b"\n"))
+		.and_then(|()| writer.flush())
+	{
+		Ok(()) => Ok(()),
+		Err(error) if error.kind() == ErrorKind::BrokenPipe => Ok(()),
+		Err(error) => Err(error.into()),
+	}
+}
